backend/requirements.txt
flask
flask-socketio
schedule
web3
neo4j
geopy
cryptography
reportlab
aiohttp
aiohttp_socks
websockets
python-dotenv
scikit-learn
numpy
networkx
ipfshttpclient